update ProgressHeader after application of blunds
It's still not very good, but slightly better than before: now the
progress is updated after the blocks are *applied* rather than after
they're *downloaded*.

Caused unfortunate ancillary changes to generator. Why? Because the
generator *applies* blocks, which is madness! It's called a generator,
it should just make the blocks... but that's another issue.